clippy: run clippy --fix

+++ src/test/utils.rs
@@ -157,7 +157,7 @@ pub fn mk_bin_contents<F: FnOnce(File) -> io::Result<()>>(
     let f = fs::OpenOptions::new()
         .write(true)
         .create(true)
-        .mode(0o666 | (libc::S_IXUSR as u32))
+        .mode(0o666 | libc::S_IXUSR)
         .open(&bin)?;
     fill_contents(f)?;
     bin.canonicalize()
